+++ models/workshops.model.js
@@ -2,7 +2,7 @@ const db = require('../utils/db.js');
 const { v4: uuidv4 } = require('uuid');
 
 module.exports = class Workshops {
-    constructor(idTaller, idCapacitacion, nombreTaller, horaEntrada, horaSalida, estatus, idUsuario) {
+    constructor(idTaller, idCapacitacion, nombreTaller, horaEntrada, horaSalida, estatus, idUsuario, horarioTaller, fecha, url) {
         this.idTaller = idTaller || uuidv4();
         this.idCapacitacion = idCapacitacion;
         this.nombreTaller = nombreTaller;
@@ -10,28 +10,30 @@ module.exports = class Workshops {
         this.horaSalida = horaSalida;
         this.estatus = estatus;
         this.idUsuario = idUsuario;
+        this.horarioTaller = horarioTaller;
+        this.fecha = fecha;
+        this.url = url;
     }
 
     async save() {
         try {
-            const query = `INSERT INTO Taller 
-                            (idTaller, 
-                            idCapacitacion, 
-                            nombreTaller, 
-                            horaEntrada, 
-                            horaSalida, 
-                            estatus, 
-                            idUsuario) 
-                            VALUES (?, ?, ?, ?, ?, ?, ?)`;
+            const query = `
+                INSERT INTO Taller 
+                (idTaller, idCapacitacion, nombreTaller, horaEntrada, horaSalida, estatus, idUsuario, HorarioTaller, Fecha, URL) 
+                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+            `;
             
             const values = [
-            this.idTaller,
-            this.idCapacitacion,
-            this.nombreTaller,
-            this.horaEntrada,
-            this.horaSalida,
-            this.estatus,
-            this.idUsuario
+                this.idTaller,
+                this.idCapacitacion,
+                this.nombreTaller,
+                this.horaEntrada,
+                this.horaSalida,
+                this.estatus,
+                this.idUsuario,
+                this.horarioTaller,
+                this.fecha,
+                this.url
             ];
 
             const result = await db.execute(query, values);
@@ -43,17 +45,21 @@ module.exports = class Workshops {
         }
     }
 
-
     static async add(tallerData) {
         try {
             const camposValidos = [
-                'idTaller', 
-                'idCapacitacion', 
-                'nombreTaller', 
-                'horaEntrada', 
-                'horaSalida', 
-                'estatus', 
-                'idUsuario'];
+                'idTaller',
+                'idCapacitacion',
+                'nombreTaller',
+                'horaEntrada',
+                'horaSalida',
+                'estatus',
+                'idUsuario',
+                'HorarioTaller',
+                'Fecha',
+                'URL'
+            ];
+
             const campos = Object.keys(tallerData).filter(key => camposValidos.includes(key));
             const placeholders = campos.map(() => '?').join(', ');
             const valores = campos.map(campo => tallerData[campo]);
@@ -68,18 +74,21 @@ module.exports = class Workshops {
         }
     }
 
-    static async update(idTaller, nombreTaller, horaEntrada, horaSalida, estatus) {
+    static async update(idTaller, nombreTaller, horaEntrada, horaSalida, estatus, horarioTaller, fecha, url) {
         try {
             const query = `
-            UPDATE Taller 
-            SET nombreTaller = ?,
-                horaEntrada = ?,
-                horaSalida = ?,
-                estatus = ?
-            WHERE idTaller = ?
+                UPDATE Taller 
+                SET nombreTaller = ?,
+                    horaEntrada = ?,
+                    horaSalida = ?,
+                    estatus = ?,
+                    HorarioTaller = ?,
+                    Fecha = ?,
+                    URL = ?
+                WHERE idTaller = ?
             `;
             
-            const params = [nombreTaller, horaEntrada, horaSalida, estatus, idTaller];
+            const params = [nombreTaller, horaEntrada, horaSalida, estatus, horarioTaller, fecha, url, idTaller];
             const result = await db.execute(query, params);
             return result;
             
@@ -87,5 +96,76 @@ module.exports = class Workshops {
             console.error('Error update():', error);
             throw error;
         }
+    }
+
+    static async getWorkshops(){
+        try{
+            const rows = await db.query
+            (`
+                SELECT t.idTaller, 
+                t.nombreTaller, 
+                t.horaEntrada, 
+                t.horaSalida,
+                t.HorarioTaller,
+                t.Fecha,
+                t.URL
+                FROM Taller t 
+                WHERE t.estatus = ?`, [1]
+            );
+            return rows;
+        } catch (err) {
+            console.error("Error fetching workshop catalog:", err);
+            throw err;
+        }
+    }
+
+    static async getOneWorkshop(id){ 
+        try{
+            const rows = await db.query(
+                `SELECT 
+	                t.nombreTaller, t.horaEntrada, t.horaSalida, c.nombreCapacitacion,
+	                u.nombre, u.apellidoPaterno
+                FROM Taller t
+                JOIN  UsuarioCapacitacion i
+	                ON t.idCapacitacion = i.idCapacitacion
+                JOIN Capacitaciones c
+	                ON i.idCapacitacion = c.idCapacitacion
+                JOIN Usuarios u
+	                ON u.idUsuario = i.idUsuario
+                WHERE t.estatus = ? AND c.estatus = ? AND t.idTaller = ?
+                `, [1, 1, id]
+            );
+            const beneficiariesRows = await db.query(
+                `SELECT  
+	                b.nombre, b.apellidoMaterno, b.apellidoPaterno, b.foto
+                FROM Beneficiarios b
+                JOIN BeneficiarioTaller e
+	                ON e.idBeneficiario = b.idBeneficiario
+                JOIN Taller t	
+	                ON e.idTaller = t.idTaller
+                WHERE t.estatus = ? AND b.estatus = ? AND t.idTaller = ?
+                `, [1, 1, id]
+            );
+            return {
+                workshop : rows,
+                beneficiaries : beneficiariesRows
+            };
+        }
+        catch(err) {
+            console.log("Error fetching one workshop", err);
+            throw err;
+        }
+        
+    }
+    static async getOneWorkshopBeneficiaries(id){ 
+        try{
+            
+            return rows;
+        }
+        catch(err) {
+            console.log("Error fetching workshop beneficiaries", err);
+            throw err;
+        }
+        
     }
 };