+++ routes/general.routes.js
@@ -9,6 +9,7 @@ const externalCollabsRoutes = require("./externalCollabs.routes");
 const supplyBatchRouter = require("./supplyBatch.routes")
 const productBatchRouter = require("./productBatch.routes")
 const productRoutes = require("./product.routes")
+const calendarRoutes = require('./calendar.routes');
 
 router.use("/beneficiary", beneficiary_routes);
 // Supplies routes
@@ -18,5 +19,6 @@ router.use("/supplyBatch", supplyBatchRouter);
 router.use("/workshop", workshopsRoutes);
 router.use("/productBatch", productBatchRouter);
 router.use("/product", productRoutes);
+router.use("/calendar", calendarRoutes);
 
 module.exports = router;